core/create: move into the subdirectory
