+++ src/core/index.ts
@@ -1,3 +1,2 @@
-export * from "./optimisticUi";
-export * from "./sync";
+export * from "./create";
 export * from "./types";
