+++ src/core/types.ts
@@ -1,3 +1,29 @@
+import NanoEvents from "nanoevents";
+import { Connectivity, NetworkAdapter } from "../network/types";
+
+export type CreateParams<Domain> = {
+  network: NetworkAdapter<Domain>;
+  onError?: (error: Error) => void;
+  window?: Window;
+};
+
+export type FullDB<Domain> = {
+  _internals: {
+    canonData: Data<Domain>;
+    debugConfig: DebugConfig;
+    events: NanoEvents<Events<Domain>>;
+    pendingTransactionCount: () => number;
+    withPendingTransactions: (data: Data<Domain>) => Data<Domain>;
+  };
+  connectivity: () => Connectivity;
+  observe: (kind: keyof Domain, id: string) => () => void;
+  update: <K extends keyof Domain>(
+    kind: K,
+    id: string,
+    delta: Delta<Domain, K>
+  ) => void;
+};
+
 export type Data<Domain> = {
   [k in keyof Domain]: {
     [id: string]: { revision: string; value: Domain[k] } | undefined;
