+++ src/index.ts
@@ -1,8 +1,7 @@
-export { Data, DebugConfig, Delta, Events } from "./core";
+export { CreateParams, Data, DebugConfig, Delta, FullDB, Events } from "./core";
 export {
   makeDummyAdapter,
   makeWsProtocolAdapter,
   Connectivity,
   NetworkAdapter
 } from "./network";
-export { CreateParams, FullDB } from "./types";
